+++ src/common.js
@@ -146,7 +146,7 @@ module.exports = {
         'block-spacing': 'error',
         'brace-style': 'error',
         'camelcase': 'error',
-        'capitalized-comments': 'error',
+        'capitalized-comments': 'ignoreConsecutiveComments',
         'comma-dangle': ['error', 'always-multiline'],
         'comma-spacing': 'error',
         'comma-style': 'error',